Private Header folder is created and qdmi_internal.h is moved
