+++ src/private/qdmi_internal.h
